Make email_address and venv_dir variables, only kill if , only restart if not Finished

+++ scripts/load_biofloat_cache_watchdog.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 # Watch the execution of a long-running load_biofloat_cache.py job
 # and restart it if its log file hasn't been updated in 10 minutes
-# or so,  Check last line for completion before restarting.
+# or so,  Check last line for completion before restarting.  The arguemnts
+# are: <log_file_name> <load_biofloat_cache.py command with all arguments>.
 #
 # Execute from cron like:
-# * */10 * * *  load_biofloat_cache_watchdog.sh /u/mccann/dev/biofloatgit/cache_age365_MR_dropna.out /u/mccann/dev/biofloatgit/scripts/load_biofloat_cache.py --age 364 --cache_dir /data/biofloat -v
+# * */10 * * *  load_biofloat_cache_watchdog.sh /u/mccann/dev/biofloatgit/cache_age364_MR_dropna.out /u/mccann/dev/biofloatgit/scripts/load_biofloat_cache.py --age 364 --cache_dir /data/biofloat -v
+
+email_address=mccann
+venv_dir=/u/mccann/dev/biofloatgit/venv-biofloat
 
 log_file=$(echo $@ | cut -d ' ' -f 1)
 shift
@@ -20,13 +24,21 @@ log_file_age=$((ct - ft))
 if [ $log_file_age -gt 600 ]
 then
     echo Paused
-    echo "$lbc_cmd  has paused.  Attempting to kill and restart." | mail -s $log_file mccann
+    echo "$lbc_cmd  has paused.  Attempting to kill and restart." | mail -s $log_file $email_address
     # Can't seem to grep for $lbc_args, grep for load_biofloat_cache.py instead
     ##pid=$(ps -ef | grep "\"$lbc_args\"" | awk -F' ' '{print $2}')
     pid=$(ps -ef | grep load_biofloat_cache.py | grep -v bash | grep -v grep | awk -F' ' '{print $2}')
-    echo killing $pid
-    kill $pid
-    echo restarting python $lbc_cmd 
-    source /u/mccann/dev/biofloatgit/venv-biofloat/bin/activate
-    python $lbc_cmd >> $log_file 2>&1 &
+    if [ ! -z $pid ]
+    then
+        echo Killing $pid
+        kill $pid
+    fi
+    last_line=$(tail -1 $log_file | cut -d ' ' -f 1)
+    if [ ! "$last_line" == "Finished" ]
+    then
+        echo Restarting python $lbc_cmd 
+        source $venv_dir/bin/activate
+        python $lbc_cmd >> $log_file 2>&1 &
+    fi
 fi
+
